correções de nomes e estrutura clean arch

+++ src/main/java/br/com/fiap/oficina/ordemservico/application/dtos/AdicionarItemPecaOrcamentoCommand.java
@@ -1,5 +1,6 @@
 package br.com.fiap.oficina.ordemservico.application.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record AdicionarItemPecaOrcamentoCommand(UUID ordemId, String itemEstoqueCodigo, double quantidade) {}
+public record AdicionarItemPecaOrcamentoCommand(UUID ordemId, String itemEstoqueCodigo, BigDecimal quantidade) {}
