+++ src/main/java/br/com/fiap/oficina/ordemservico/application/dtos/AdicionarItemServicoOrcamentoCommand.java
@@ -1,5 +1,6 @@
 package br.com.fiap.oficina.ordemservico.application.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record AdicionarItemServicoOrcamentoCommand(UUID ordemId, String servicoCodigo, double quantidade) {}
+public record AdicionarItemServicoOrcamentoCommand(UUID ordemId, String servicoCodigo, BigDecimal quantidade) {}
