+++ src/main/java/br/com/fiap/oficina/ordemservico/interfaceadapters/presenters/rest/response/OrcamentoItemServicoResponse.java
@@ -1,8 +1,9 @@
 package br.com.fiap.oficina.ordemservico.interfaceadapters.presenters.rest.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrcamentoItemServicoResponse(
         UUID servicoId,
-        Double quantidade
+        BigDecimal quantidade
 ) {}
